Fix sonarcloud issues
