+++ exp/templates/exp/support.html
@@ -68,9 +68,6 @@
         <iframe src="https://docs.google.com/forms/d/e/1FAIpQLScy0DBdFNvDPsd-YkqiDkykoUmpjEfZCObOBefW7w-ekw5v0w/viewform?embedded=true"
                 width="640"
                 height="2100"
-                frameborder="0"
-                marginheight="0"
-                marginwidth="0"
                 title="Lookit - Citations & Impact">Loading…</iframe>
         <h3>Write Lookit into your grants</h3>
         <p>
