+++ locale/de/LC_MESSAGES/django.po
@@ -1199,7 +1199,7 @@ msgid ""
 "Lookit is available to all developmental researchers, and the code is "
 "entirely open- To learn more about running your own studies or contributing, "
 "please check out the <a href='https://lookit.readthedocs.io/en/develop' "
-"target='_blank'>documentation</a>."
+"target="_blank">documentation</a>."
 msgstr ""
 
 #: web/templates/frontpages/contact.html:21
@@ -3177,7 +3177,7 @@ msgstr ""
 #~ msgid ""
 #~ "For even more ways to participate in science with your kids at home, "
 #~ "check out the newly launched site <a href='https://childrenhelpingscience."
-#~ "com/' target='_blank'>Children Helping Science"
+#~ "com/' target="_blank">Children Helping Science"
 #~ msgstr ""
 #~ "Suchen Sie nach weiteren Möglichkeiten, um von zu Hause aus zur Forschung "
 #~ "beizutragen? Weitere Studien finden Sie auf <a href=\"https://"
