+++ locale/it/LC_MESSAGES/django.po
@@ -1199,7 +1199,7 @@ msgid ""
 "Lookit is available to all developmental researchers, and the code is "
 "entirely open- To learn more about running your own studies or contributing, "
 "please check out the <a href='https://lookit.readthedocs.io/en/develop' "
-"target='_blank'>documentation</a>."
+"target="_blank">documentation</a>."
 msgstr ""
 
 #: web/templates/frontpages/contact.html:21
