+++ locale/nl/LC_MESSAGES/django.po
@@ -1204,7 +1204,7 @@ msgid ""
 "Lookit is available to all developmental researchers, and the code is "
 "entirely open- To learn more about running your own studies or contributing, "
 "please check out the <a href='https://lookit.readthedocs.io/en/develop' "
-"target='_blank'>documentation</a>."
+"target="_blank">documentation</a>."
 msgstr ""
 
 #: web/templates/frontpages/contact.html:21
@@ -3157,7 +3157,7 @@ msgstr ""
 #~ msgid ""
 #~ "For even more ways to participate in science with your kids at home, "
 #~ "check out the newly launched site <a href='https://childrenhelpingscience."
-#~ "com/' target='_blank'>Children Helping Science"
+#~ "com/' target="_blank">Children Helping Science"
 #~ msgstr ""
 #~ "Op zoek naar meer manieren om bij te dragen aan onderzoek vanuit huis? "
 #~ "Bekijk <a href=\"https://childrenhelpingscience.com/\" target=\"_blank\" "
