+++ locale/pt_BR/LC_MESSAGES/django.po
@@ -1198,7 +1198,7 @@ msgid ""
 "Lookit is available to all developmental researchers, and the code is "
 "entirely open- To learn more about running your own studies or contributing, "
 "please check out the <a href='https://lookit.readthedocs.io/en/develop' "
-"target='_blank'>documentation</a>."
+"target="_blank">documentation</a>."
 msgstr ""
 
 #: web/templates/frontpages/contact.html:21
@@ -3173,7 +3173,7 @@ msgstr ""
 #~ msgid ""
 #~ "For even more ways to participate in science with your kids at home, "
 #~ "check out the newly launched site <a href='https://childrenhelpingscience."
-#~ "com/' target='_blank'>Children Helping Science"
+#~ "com/' target="_blank">Children Helping Science"
 #~ msgstr ""
 #~ "Procurando mais maneiras de contribuir sem sair de casa? Confira <a href="
 #~ "\"https://childrenhelpingscience.com/\" target=\"_blank\" rel="
