+++ studies/templates/studies/study_attachments.html
@@ -79,20 +79,21 @@
         <div class="row mt-md">
             <div class="col-xs-12">
                 <div class="table-responsive">
-                    <table class="table table-striped">
+                    <table class="study-attachments table table-striped">
+                        <caption>Study Attachments</caption>
                         <thead>
                             <tr>
                                 <th scope="col">
                                     Name
                                     <a class="pl-sm"
-                                       href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='full_name' %}"><i class="fa fa-chevron-up" role="button"></i></a>
-                                    <a href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='-full_name' %}"><i class="fa fa-chevron-down" role="button"></i></a>
+                                       href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='full_name' %}"><em class="fa fa-chevron-up" role="button"></em></a>
+                                    <a href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='-full_name' %}"><em class="fa fa-chevron-down" role="button"></em></a>
                                 </th>
                                 <th scope="col">
                                     Date
                                     <a class="pl-sm"
-                                       href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='created_at' %}"><i class="fa fa-chevron-up" role="button"></i></a>
-                                    <a href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='-created_at' %}"><i class="fa fa-chevron-down" role="button"></i></a>
+                                       href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='created_at' %}"><em class="fa fa-chevron-up" role="button"></em></a>
+                                    <a href="{% url 'exp:study-attachments' study.id %}?{% query_transform request sort='-created_at' %}"><em class="fa fa-chevron-down" role="button"></em></a>
                                 </th>
                                 <th scope="col"></th>
                             </tr>
