+++ web/templates/frontpages/home.html
@@ -50,11 +50,3 @@
         </div>
     </div>
 {% endblock content %}
-{% block footerlinks %}
-    <li>
-        <a href="{% url 'web:privacy' %}">{% trans "Privacy" %}</a>
-    </li>
-    <li>
-        <a href="{% url 'web:contact' %}">{% trans "Contact us" %}</a>
-    </li>
-{% endblock footerlinks %}
