+++ web/templates/frontpages/resources.html
@@ -35,7 +35,7 @@
                                             </select>
                                         </div>
                                         <div class="col-md-8">
-                                            <b class="selected-state"></b>
+                                            <strong class="selected-state"></strong>
                                             <p id="nothing-to-show">
                                             </p>
                                             <ul id="lab-list">
@@ -48,73 +48,73 @@
                             <div class="resources-item">
                                 <h3>Other ways to participate from home</h3>
                                 <p>
-                                    Visit <a href='https://childrenhelpingscience.com/' target='_blank'>Children Helping
-                                Science</a> for
-                                an
-                                extensive list of other opportunities to participate in research about child
-                                development!
+                                    Visit
+                                    <a href="https://childrenhelpingscience.com/"
+                                       target="_blank"
+                                       rel="noopener">Children Helping Science</a>
+                                    for an extensive list of other opportunities to participate in research about child development!
+                                </p>
+                                <p>
+                                To contribute to other scientific research as a family, check out <a href="https://scistarter.org/">SciStarter</a>, which has many citizen
+                                science projects
+                                suitable for elementary school children and up!
                             </p>
+                        </div>
+                        <hr />
+                        <div class="resources-item">
+                            <h3>Activities to try at home</h3>
                             <p>
-                            To contribute to other scientific research as a family, check out <a href="https://scistarter.org/">SciStarter</a>, which has many citizen
-                            science projects
-                            suitable for elementary school children and up!
+                                Want to learn more about cognitive development? Here are some activities that may give
+                                you some
+                                insight
+                                into your own child's developing mind. Instead of studies our lab is running, these are
+                                "at home
+                                labs"
+                                for parents to try on their own--but please feel free to contact us with any questions!
+                            </p>
+                            <h4>Learning about other minds: your child's developing theory of mind</h4>
+                            <iframe allowfullscreen
+                                    height="315"
+                                    src="https://www.youtube.com/embed/uxyVYATX9-M"
+                                    width="560"
+                                    title="YouTube Video"></iframe>
+                            <p>
+                                <strong>Age range</strong>: 2.5 to 5 years
                         </p>
-                    </div>
-                    <hr />
-                    <div class="resources-item">
-                        <h3>Activities to try at home</h3>
                         <p>
-                            Want to learn more about cognitive development? Here are some activities that may give
-                            you some
-                            insight
-                            into your own child's developing mind. Instead of studies our lab is running, these are
-                            "at home
-                            labs"
-                            for parents to try on their own--but please feel free to contact us with any questions!
-                        </p>
-                        <h4>Learning about other minds: your child's developing theory of mind</h4>
-                        <iframe allowfullscreen
-                                frameborder="0"
-                                height="315"
-                                src="https://www.youtube.com/embed/uxyVYATX9-M"
-                                width="560"></iframe>
-                        <p>
-                            <strong>Age range</strong>: 2.5 to 5 years
+                            <strong>What you'll need</strong>: For the "Maxi and the chocolate" story, any props
+                        you'd like (you
+                        can
+                        try drawing a picture or acting out the story). For the word-learning task, two
+                        containers to hold
+                        two
+                        objects your child doesn't know a name for (weird kitchen tools, bike parts, etc.).
                     </p>
                     <p>
-                        <strong>What you'll need</strong>: For the "Maxi and the chocolate" story, any props
-                    you'd like (you
-                    can
-                    try drawing a picture or acting out the story). For the word-learning task, two
-                    containers to hold
-                    two
-                    objects your child doesn't know a name for (weird kitchen tools, bike parts, etc.).
+                        In this lab, you'll see how your child thinks about what other people are thinking.
+                        Children under
+                        about
+                        four years old tend to have trouble expressing what's going on when someone else knows
+                        less than
+                        they
+                        do. They will often insist that everyone knows what they themselves know!
+                    </p>
+                    <h4>Learning to count: measuring your child's N-knower level</h4>
+                    <iframe allowfullscreen
+                            height="315"
+                            src="https://www.youtube.com/embed/i0q6H8MRXo8"
+                            width="560"
+                            title="YouTube Video">></iframe>
+                    <p>
+                        <strong>Age range</strong>: 1 to 5 years
                 </p>
                 <p>
-                    In this lab, you'll see how your child thinks about what other people are thinking.
-                    Children under
-                    about
-                    four years old tend to have trouble expressing what's going on when someone else knows
-                    less than
-                    they
-                    do. They will often insist that everyone knows what they themselves know!
-                </p>
-                <h4>Learning to count: measuring your child's N-knower level</h4>
-                <iframe allowfullscreen
-                        frameborder="0"
-                        height="315"
-                        src="https://www.youtube.com/embed/i0q6H8MRXo8"
-                        width="560"></iframe>
-                <p>
-                    <strong>Age range</strong>: 1 to 5 years
+                    <strong>What you'll need</strong>: At least ten small objects your child can pick up,
+                like pegs or
+                Cheerios
             </p>
-            <p>
-                <strong>What you'll need</strong>: At least ten small objects your child can pick up,
-            like pegs or
-            Cheerios
-        </p>
+        </div>
     </div>
-</div>
 </div>
 </div>
 </div>
