+++ web/templates/frontpages/termsofuse.html
@@ -473,9 +473,11 @@
                 <p class="last-update">
                     Last updated October 13, 2021
                     <a href="https://github.com/lookit/research-resources/commits/master/lookit%20static%20pages/termsofuse.html"
-                       target="_blank">[Changelog]</a>
+                       target="_blank"
+                       rel="noopener">[Changelog]</a>
                     <a href="https://mailman.mit.edu/mailman/listinfo/lookit-terms-notifications"
-                       target="_blank">[Change notifications]</a>
+                       target="_blank"
+                       rel="noopener">[Change notifications]</a>
                 </p>
             </div>
         </div>
