+++ web/templates/web/participant-signup.html
@@ -31,7 +31,7 @@
                     {% bootstrap_form form %}
                     {% buttons %}
                     <p>
-                        {% trans "By clicking 'Create Account', I agree that I have read and accepted the " %} <a href="{% url 'web:privacy' %}" target="_blank">{% trans "Privacy Statement" %}.</a>
+                        {% trans "By clicking 'Create Account', I agree that I have read and accepted the " %} <a href="{% url 'web:privacy' %}" target="_blank" rel="noopener">{% trans "Privacy Statement" %}.</a>
                     </p>
                     <div class="pull-right">
                         <button type="submit" class="btn btn-success">{% bootstrap_icon "plus" %} {% trans "Create Account" %}</button>
